Added a dialog to delete the hero build

+++ app/src/main/java/com/example/tanorami/builds_hero_from_users/BuildsHeroListFragment.kt
@@ -99,7 +99,7 @@ class BuildsHeroListFragment :
 
     private fun setupAdapter() {
         mAdapter = BuildsHeroListAdapter(object : BuildsHeroListListener {
-            override fun onClick(idBuild: Int) {
+            override fun onClick(idBuild: Long) {
                 findNavController().navigate(R.id.viewingUsersBuildFragment, ViewingUsersBuildFragment.newInstance(idBuild))
             }
         })
