+++ app/src/main/java/com/example/tanorami/builds_hero_from_users/adapter/BuildsHeroListListener.kt
@@ -1,5 +1,5 @@
 package com.example.tanorami.builds_hero_from_users.adapter
 
 interface BuildsHeroListListener {
-    fun onClick(idBuild: Int)
+    fun onClick(idBuild: Long)
 }
